isolated the create icon/font/icons in dedicated module
