+++ packages/pymodaq_gui/src/pymodaq_gui/examples/material_icon_ex.py
@@ -1,6 +1,6 @@
 from qtpy import QtGui, QtWidgets
 from pymodaq_gui.resources.material_icons import MaterialIcon
-from pymodaq_gui.managers.action_manager import resource_path_exists
+from pymodaq_gui.utils.styling import resource_path_exists
 from pymodaq_gui.utils.custom_app import CustomApp
 from pymodaq_utils.config import Config
